requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
pandas==2.2.2
fpdf2==2.6.1
openpyxl==3.1.2
